refactor(userbodymetrics): move create use case to the own package
