+++ src/main/java/com/ironcore/application/userbodymetrics/create/CreateUserBodyMetricsUseCase.java
@@ -1,12 +1,11 @@
-package com.ironcore.application.userbodymetrics;
+package com.ironcore.application.userbodymetrics.create;
 
 import com.ironcore.application.exception.OperationNotAllowedException;
 import com.ironcore.application.exception.ResourceNotFoundException;
 import com.ironcore.application.exception.UserInactiveException;
 import com.ironcore.application.logging.audit.port.AuditLogPublisher;
+import com.ironcore.application.userbodymetrics.UserBodyMetricsAuditData;
 import com.ironcore.application.userbodymetrics.component.BodyFatPercentageCalculator;
-import com.ironcore.application.userbodymetrics.create.CreateUserBodyMetricsCommand;
-import com.ironcore.application.userbodymetrics.create.CreateUserBodyMetricsResult;
 import com.ironcore.domain.logging.audit.enums.AuditActionType;
 import com.ironcore.domain.logging.audit.enums.AuditTargetType;
 import com.ironcore.domain.user.model.User;
